Don't block on read to redraw
When the shape of a terminal is changes, redraw() recives a WINCH
signal to redraw its content. Unfortunately, it's not the case
when read(1) is invoked. The changes are here to work around the
issue to allow the script redraw its content ten times per second.

Also I finally fixed cursor misplacement issue. For real at this time.